refactor: 노선의 구간 관련 기능 분리
